+++ branches/nhd2-exp/src/system/ytparser.cpp
@@ -53,7 +53,7 @@ void cYTVideoInfo::Dump()
 	printf("title: %s\n", title.c_str());
 	printf("duration: %d\n", duration);
 	//printf("description: %s\n", description.c_str());
-	printf("urls: %d\n", formats.size());
+	printf("urls: %d\n", (int)formats.size());
 	for (yt_urlmap_iterator_t it = formats.begin(); it != formats.end(); ++it) {
 		printf("format %d type [%s] url %s\n", it->first, it->second.type.c_str(), it->second.GetUrl().c_str());
 	}
@@ -115,7 +115,7 @@ bool cYTFeedParser::getUrl(std::string &url, std::string &answer)
 
 	curl_easy_cleanup(curl_handle);
 
-	printf("http: res %d size %d\n", httpres, answer.size());
+	printf("http: res %d size %d\n", httpres, (int)answer.size());
 
 	if (httpres != 0 || answer.empty()) {
 		printf("error: %s\n", cerror);
@@ -582,7 +582,7 @@ bool cYTFeedParser::GetVideoUrls()
 
 void cYTFeedParser::Cleanup(bool delete_thumbnails)
 {
-	printf("cYTFeedParser::Cleanup: %d videos\n", videos.size());
+	printf("cYTFeedParser::Cleanup: %d videos\n", (int)videos.size());
 	if (delete_thumbnails) {
 		for (unsigned i = 0; i < videos.size(); i++) {
 			unlink(videos[i].tfile.c_str());
@@ -596,7 +596,7 @@ void cYTFeedParser::Cleanup(bool delete_thumbnails)
 
 void cYTFeedParser::Dump()
 {
-	printf("feed: %d videos\n", videos.size());
+	printf("feed: %d videos\n", (int)videos.size());
 	for (unsigned i = 0; i < videos.size(); i++)
 		videos[i].Dump();
 }
